# swagger.yaml
openapi: 3.0.0
info:
  title: Learn Express and Swagger.
  description: A course segment for writing docs.
  version: 1.1.0
  contact:
    email: manish@gmail.com
    url: "imvmanish.com"

# servers:
#   - url: "https://localhost:3000/api/v1"
#     description: for local host - secure
#   - url: "http://localhost:3000/api/v1"
#     description: for local host

servers:
  - url: "{protocol}://localhost:3000/api/{version}"
    description: for local host 
    variables:
      version:
        enum: 
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - https
          - http
        default: http

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: Token


# paths:
#   /lco:
#     get:
#     post:
#   /api:
#     put:
#     get:

paths:
  /lco:
    get:
      tags:
      - String
      summary: returns a hello world message
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: string
  /lcoobject:
    get:
      tags:
      - Object
      summary: returns an object
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: number             
        '400':  
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string             
        '500':  
          description: Internal Server Error

  /lcoarray:
    get:
      tags:
      - Array
      summary: returns an array of courses
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    price:
                      type: number             
        '400':  
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string             
        '500':  
          description: Internal Server Error
  /course/{id}:
    get:
      tags:
      - Object
      summary: returns a course of given id
      parameters:
        - in: path
          name: id
          default: 1
          schema:
            type: integer
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: number             
        '400':  
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string             
        '500':  
          description: Internal Server Error

  /addCourse:
    post:
      tags:
      - Boolean
      summary: adds a course
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                id:
                  type: number
                name:
                  type: string
                price:
                  type: number
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: boolean            
        '400':  
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string             
        '500':  
          description: Internal Server Error
  /search:
    get:
      tags:
      - Object
      summary: search something
      parameters:
        - name: location
          in: query
          schema:
            type: string
            enum: [bangalore,delhi,noida]
          required: true
        - name: device
          in: query
          schema:
            type: string
            enum: [web,mobile]
          required: true
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string            
        '400':  
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string             
        '500':  
          description: Internal Server Error
  /upload:
    post:
      tags:
      - Images
      summary: uploading course image
      consumes:
        - multipart/form-data
      produces:
        - application/json
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: 
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':  
          description: Successful OK.
          content:
            application/json:
              schema: 
                type: string           
        '400':  
          description: Bad Request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string             
        '500':  
          description: Internal Server Error